+++ src/lib_wd1793/wd1793_crc_utils.h
@@ -3,8 +3,8 @@
 
 #include "wd1793_defs.h"
 
-extern unsigned wd1793_crc(uint8_t *ptr, unsigned size);
-extern unsigned short wd1793_crc16(uint8_t *buf, unsigned size);
+extern uint16_t wd1793_crc(uint8_t *ptr, unsigned size);
+extern uint16_t wd1793_crc16(uint8_t *buf, unsigned size);
 extern void wd1793_crc32(int &crc, uint8_t *buf, unsigned len);
 
 #endif
